Feat: new exception on neighbor api

+++ src/APIs/neighbors/neighbors.service.ts
@@ -2,6 +2,7 @@ import { ConflictException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Neighbor } from './entities/neighbor.entity';
 import { DataSource, Repository } from 'typeorm';
+import { findSourceMap } from 'module';
 
 @Injectable()
 export class NeighborsService {
@@ -11,6 +12,12 @@ export class NeighborsService {
     private readonly dataSource: DataSource,
   ) {}
 
+  isSame({ from_user, to_user }): boolean {
+    if (from_user == to_user) {
+      return true;
+    }
+    return false;
+  }
   async isExist({ from_user, to_user }): Promise<boolean> {
     console.log(from_user, to_user);
     const neighbor = await this.neighborsRepository.findOne({
@@ -30,6 +37,9 @@ export class NeighborsService {
     if (isExist) {
       throw new ConflictException('already exists');
     }
+    if (this.isSame({ from_user, to_user })) {
+      throw new ConflictException('you cannot follow yourself!');
+    }
     const neighbor = await this.neighborsRepository.save({
       from_user,
       to_user,
@@ -41,6 +51,9 @@ export class NeighborsService {
     const isExist = await this.isExist({ from_user, to_user });
     if (!isExist) {
       throw new ConflictException('no data exists');
+    }
+    if (this.isSame({ from_user, to_user })) {
+      throw new ConflictException('you cannot unfollow yourself!');
     }
     return this.neighborsRepository.delete({ from_user, to_user });
   }
